+++ src/app/app-routing.module.ts
@@ -4,7 +4,7 @@ import type { Routes } from '@angular/router';
 import { RouterModule } from '@angular/router';
 import { LoginComponent } from './login/login.component';
 import { ProgramComponent } from './program/view/program.component';
-import { VotingComponent } from './voting/voting.component';
+import { VotingComponent } from './voting/view/voting.component';
 import { AddEventComponent } from './program/add/add-event/add-event.component';
 import { EditEventComponent } from './program/edit/edit-event/edit-event.component';
