+++ src/app/services/voting/voting-result.service.spec.ts
