+++ src/app/voting/add-edit/voting-add-edit.component.spec.ts
@@ -7,7 +7,7 @@ import {
   MAT_DIALOG_DATA,
 } from '@angular/material/dialog';
 import { EventService } from '../../services/event.service';
-import { VotingService } from '../../services/voting.service';
+import { VotingService } from '../../services/voting/voting.service';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 
 describe('VotingAddEditComponent', () => {
