+++ src/app/voting/view/voting-wrapper/voting-element/voting-card/voting-card.component.spec.ts
