+++ src/app/voting/view/voting-wrapper/voting-element/voting-element.component.html
