+++ src/app/voting/view/voting-wrapper/voting-wrapper.component.spec.ts
