+++ src/app/voting/view/voting.component.css
