+++ src/app/voting/view/voting.component.html
