+++ src/app/voting/view/voting.component.spec.ts
